exported textures are ordered as they appear
so initially we would just go to the folder and package it into a WAD.
TexMex on the other hand also follows the order of the textures as they
appear, which means we can organize the WAD itself. This commit makes
qthon do the same thing.
However it's also kind of pointless, because I was expecting trenchbroom
to also follow the order of textures XD