implemented water animation preview
literally the worst possible implementation that a man can think of. I
found online some code that pretty much recreates the quake 1 water
animation in html+js. I pretty much took that, simplified it, and wrote
a python module that loads it, replaces the image to the one we want to
view, start the server AT THE TEMP DIR, and we display that.

I would also like to credit the person that wrote that html+js code, but
their website is broken at the moment. Will include that info in the
about and readme later