added defullbright prompt window
now whenever we try to defullbright texture(s), we will get a window
that asks if we want to include the `-dfb` suffix or not, with a
checkbox to always ask for that prompt. If disabled the prompt, you can
still disable/enable the defullbright overwriting in the preferences
window

I recently had a problem where I had to defullbright a lot of textures
(like 50 or so), so I had to manually delete duplicates with fullbright
pixels. I didn't bother removig the suffix from texture names, couldn't
be bothered with that